requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
asgi-logger==0.1.0
asgiref==3.6.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
bintrees==2.2.0
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
pandas==2.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.13
starlette==0.26.1
typing-extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
yarl==1.9.2
zipp==3.15.0
